Adicionando classes DTO, paginacao e realizando ajustes me services e controllers

+++ src/main/java/com/example/bot/entities/Classe.java
@@ -17,10 +17,10 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 @Table(name="tb_classe")
 public class Classe implements Serializable{
 	private static final long serialVersionUID = 1L;
-
+	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 	
@@ -30,16 +30,16 @@ public class Classe implements Serializable{
 	
 	public Classe() {}
 
-	public Classe(Long id, String nome) {
+	public Classe(Integer id, String nome) {
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
