+++ src/main/java/com/example/bot/repositories/SemanticoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.example.bot.entities.Semantico;
 
-public interface SemanticoRepository extends JpaRepository<Semantico, Long> {
+public interface SemanticoRepository extends JpaRepository<Semantico, Integer> {
 	
 	@Query(value="SELECT * FROM TB_SEMANTICO WHERE MENSAGEM_FORMATADA = :mesagemFormatada", nativeQuery = true)
 	Semantico findByMensagemFormatada(@Param("mesagemFormatada") String mensagem);
